fix: added console to test
